refactor: モックモジュール名のリネーム
